+++ CHANGELOG.md
@@ -10,7 +10,7 @@ Check [Keep a Changelog](http://keepachangelog.com/) for recommendations on how 
 
 ## v0.2.0
 - Changed "Generating seed script boilerplate" notification message to notification with progress
-- Refactored underlying code to be more agnostic with connection provider
+- Refactored underlying code to be more agnostic with connection provider (e.g. MSSQL, etc.)
 
 ## v0.1.0 - 2023-09-10
 - Initial pre-release